classifier/requirements.txt
flask
torch
transformers
pandas
scikit-learn
flask
pandas
sentence-transformers
faiss-cpu
numpy
